# Conflicts:
#	src/main/java/top/rslly/iot/controllers/Tool.java
#	src/main/java/top/rslly/iot/services/thingsModel/ProductServiceImpl.java
#	web/src/i18n/lang/en/sidebar.js

+++ src/main/java/top/rslly/iot/services/thingsModel/ProductServiceImpl.java
@@ -78,6 +78,8 @@ public class ProductServiceImpl implements ProductService {
   private KnowledgeGraphicAttributeRepository knowledgeGraphicAttributeRepository;
   @Resource
   private ProductLlmModelRepository productLlmModelRepository;
+  @Resource
+  private ProductSkillsRepository productSkillsRepository;
 
 
   @Override
@@ -156,6 +158,7 @@ public class ProductServiceImpl implements ProductService {
         if (userList.isEmpty()) {
           return ResultTool.fail(ResultCode.COMMON_FAIL);
         }
+        wxProductBindEntity.setAppid(userList.get(0).getAppid());
         wxProductBindEntity.setOpenid(userList.get(0).getOpenid());
         wxProductBindEntity.setProductId(productEntity1.getId());
         wxProductBindRepository.save(wxProductBindEntity);
@@ -207,6 +210,8 @@ public class ProductServiceImpl implements ProductService {
         knowledgeGraphicNodeRepository.findAllByProductId(id);
     List<ProductLlmModelEntity> productLlmModelEntityList =
         productLlmModelRepository.findAllByProductId(id);
+    List<ProductSkillsEntity> productSkillsEntityList =
+        productSkillsRepository.findAllByProductId(id);
     boolean p1 = productModelEntityList.isEmpty();
     boolean p2 = wxProductBindEntityList.isEmpty();
     boolean p3 = otaEntityList.isEmpty();
@@ -226,7 +231,8 @@ public class ProductServiceImpl implements ProductService {
       }
       knowledgeGraphicNodeRepository.deleteAllByProductId(id);
     }
-    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9 && p10 && p11) {
+    boolean p12 = productSkillsEntityList.isEmpty();
+    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9 && p10 && p11 && p12) {
       List<ProductEntity> result = productRepository.deleteById(id);
       if (result.isEmpty())
         return ResultTool.fail(ResultCode.PARAM_NOT_VALID);
